experimentation-governance/requirements.txt
streamlit>=1.22.0
pandas>=1.5.3
polars>=0.19.2
pyyaml>=6.0
PyGithub>=2.1.1
requests>=2.28.0
python-dotenv>=1.0.0
github3.py>=4.0.1
PyJWT>=2.6.0
cryptography>=40.0.0
streamlit-oauth>=0.1.0
streamlit-ace>=0.1.1
beyondclient
# The code will fall back to using requests if beyondclient is not available
